==> analysis/01-PUMA_feature_transform.ipynb <==
import numpy as np
import pandas as pd
import scipy.linalg

# Load features from population.csv file
# This file is available on Zenodo entry with the dataset
population_df = pd.read_csv('../misc/population_normalized.csv.gz')

#Load broad compound ids which are used in the current PUMA experiments
broad_ids_df = pd.read_csv('../misc/broad_ids.txt')
broad_ids = broad_ids_df.broad_ids.to_list()
len(broad_ids)

# Filter population.csv, leave only compounds from PUMA experiment + DMSO
population_df_filtered = population_df[ (population_df["Metadata_broad_sample"] == "DMSO") | (population_df["Metadata_pert_id"].isin(broad_ids)) ].reset_index(drop=True).copy()
population_df_filtered

#Get lists of feature column names
feature_columns = population_df_filtered.columns[20:].tolist()
nan_columns = []
for i in feature_columns:
    if population_df_filtered[i].isnull().values.any():
        nan_columns.append(i)

feature_columns = list(set(feature_columns) - set(nan_columns))

class SpheringNormalizer(object):
    def __init__(self, controls):
        REG_PARAM = 10**np.log(1/controls.shape[0])
        # Sphering transform on population level data
        self.mu = controls.mean()
        self.sphering_transform(controls - self.mu, REG_PARAM, rotate=True)
        print(self.mu.shape, self.W.shape)
        
    def sphering_transform(self, X, lambda_, rotate=True):
        C = (1/X.shape[0]) * np.dot(X.T, X)
        s, V = scipy.linalg.eigh(C)
        D = np.diag( 1. / np.sqrt(s + lambda_) )
        W = np.dot(V, D)
        if rotate:
            W = np.dot(W, V.T)
        self.W = W

    def normalize(self, X):
        return np.dot(X - self.mu, self.W)

#Start sphering
sN = SpheringNormalizer(population_df_filtered.loc[population_df_filtered["Metadata_broad_sample"] == "DMSO", feature_columns])
sD = sN.normalize(population_df_filtered[feature_columns])

#Replace original feature values with feature values after sphering
population_df_filtered[feature_columns] = sD

#Mean aggregation
aggregated_sphered = population_df_filtered[['Metadata_broad_sample', 'Metadata_Plate_Map_Name', 'Metadata_pert_id'] + feature_columns].copy()
aggregated_sphered = aggregated_sphered.groupby("Metadata_broad_sample").mean().reset_index()

aggregated_sphered_np = aggregated_sphered[aggregated_sphered['Metadata_broad_sample'] != 'DMSO'].to_numpy()

# Save features. Those should be preprocessed later (sorted in the same way as in other experiments, remove first column)
np.savez('../feature_data/aggregated_sphered_morphology_features_norm', features=aggregated_sphered_np)
