==> analysis/04-late_fusion.ipynb <==
import numpy as np
import pandas as pd

Loop over splits, the prediction files per split can be found on Zenodo (dataset). 

The files produced by this notebook are also available on Zenodo (dataset).

for i in range(0,5):
    pred_folder = '../predictions/chemical_cv{}test/predictions/'.format(i)
    predictions_file_cs = pred_folder + 'predictions_cp_es_op.csv'
    predictions_cs = pd.read_csv(predictions_file_cs)
    cs_header = predictions_cs.columns
    predictions_cs_np = predictions_cs.to_numpy()[:, 1:].astype('float')
    cs_compounds = list(predictions_cs.to_numpy()[:, 0])

    predictions_file_mobc = pred_folder + 'predictions_mobc_es_op.csv'
    predictions_mobc = pd.read_csv(predictions_file_mobc)
    mobc_header = predictions_mobc.columns
    predictions_mobc_np = predictions_mobc.to_numpy()[:, 1:].astype('float')
    mobc_compounds = list(predictions_mobc.to_numpy()[:, 0])

    predictions_file_ge = pred_folder + 'predictions_ge_es_op.csv'
    predictions_ge = pd.read_csv(predictions_file_ge)
    ge_header = predictions_ge.columns
    predictions_ge_np = predictions_ge.to_numpy()[:, 1:].astype('float')
    ge_compounds = list(predictions_ge.to_numpy()[:, 0])
    ge_compounds_ = predictions_ge['smiles']
    
    
    ge_mobc = np.zeros(predictions_ge_np.shape).astype('float')
    cs_mobc = np.zeros(predictions_ge_np.shape).astype('float')
    cs_ge = np.zeros(predictions_ge_np.shape).astype('float')
    cs_ge_mobc = np.zeros(predictions_ge_np.shape).astype('float')
    
    
    if (list(ge_header) == list(cs_header) and list(cs_header) == list(mobc_header)):
        if (predictions_cs_np.shape == predictions_ge_np.shape and predictions_cs_np.shape == predictions_mobc_np.shape):
            if mobc_compounds == ge_compounds and ge_compounds == cs_compounds:
                for i in range(predictions_ge_np.shape[0]):
                    for j in range(predictions_ge_np.shape[1]):
                        ge_mobc[i,j] = max(predictions_ge_np[i,j], predictions_mobc_np[i,j])
                        cs_mobc[i,j] = max(predictions_cs_np[i,j], predictions_mobc_np[i,j])
                        cs_ge[i,j] = max(predictions_cs_np[i,j], predictions_ge_np[i,j])
                        cs_ge_mobc[i,j] = max(predictions_cs_np[i,j], predictions_ge_np[i,j], predictions_mobc_np[i,j])
        
    
    
    ge_mobc_df = pd.DataFrame(data = ge_mobc)
    cs_mobc_df = pd.DataFrame(data = cs_mobc)
    cs_ge_df = pd.DataFrame(data = cs_ge) 
    cs_ge_mobc_df = pd.DataFrame(data = cs_ge_mobc)
    
    ge_mobc_df = pd.concat([ge_compounds_, ge_mobc_df],axis=1)
    cs_mobc_df = pd.concat([ge_compounds_, cs_mobc_df],axis=1)
    cs_ge_df = pd.concat([ge_compounds_, cs_ge_df],axis=1)
    cs_ge_mobc_df = pd.concat([ge_compounds_, cs_ge_mobc_df],axis=1)
    
    ge_mobc_df.columns = ge_header
    cs_mobc_df.columns = ge_header
    cs_ge_df.columns = ge_header
    cs_ge_mobc_df.columns = ge_header
    
    ge_mobc_df.to_csv(pred_folder + 'late_fusion_ge_mobc.csv', index=False)
    cs_mobc_df.to_csv(pred_folder + 'late_fusion_cs_mobc.csv', index=False)
    cs_ge_df.to_csv(pred_folder + 'late_fusion_cs_ge.csv', index=False)
    cs_ge_mobc_df.to_csv(pred_folder + 'late_fusion_cs_ge_mobc.csv', index=False)
