==> analysis/09-Venn_diagrams.ipynb <==
import pandas as pd
from matplotlib import pyplot as plt
from venn import venn

# Generate Venn diagram for single modalities and AUC 0.9
df_dataset = pd.read_csv('../predictions/scaffold_median_AUC.csv')
cs = df_dataset[(df_dataset['descriptor'] == 'cp_es_op')][(df_dataset['auc_90'] == True)]
ge = df_dataset[(df_dataset['descriptor'] == 'ge_es_op')][(df_dataset['auc_90'] == True)]
mo = df_dataset[(df_dataset['descriptor'] == 'mobc_es_op')][(df_dataset['auc_90'] == True)] 

plt.figure(figsize=(4,4))
set1 = set(cs['assay_id'])
set2 = set(ge['assay_id'])
set3 = set(mo['assay_id'])

# Adjust circle sizes in Illustrator
plt.figure(figsize=(9,9))
dict_set = {'CS': set1, 'GE':set2, 'MO-BC':set3}
cmaps = ["cool", list("rgb"), "plasma", "viridis", "Set1"]
plt.figure(figsize=(9,9))
cmap = ["#F1C232","#6FA8DC","#93C47C"]
venn(dict_set, cmap=cmap, fontsize=16, legend_loc=None)
plt.savefig("../plots/venn_single_modalities_90.svg")

# Generate Venn diagram for combined modalities and AUC 0.9
df_dataset = pd.read_csv('../predictions/scaffold_median_AUC.csv')
csge = df_dataset[(df_dataset['descriptor'] == 'late_fusion_cs_ge')][(df_dataset['auc_90'] == True)]
gemo = df_dataset[(df_dataset['descriptor'] == 'late_fusion_ge_mobc')][(df_dataset['auc_90'] == True)]
csmo = df_dataset[(df_dataset['descriptor'] == 'late_fusion_cs_mobc')][(df_dataset['auc_90'] == True)]
csgemo = df_dataset[(df_dataset['descriptor'] == 'late_fusion_cs_ge_mobc')][(df_dataset['auc_90'] == True)] 

plt.figure(figsize=(4,4))
set1 = set(csge['assay_id'])
set2 = set(gemo['assay_id'])
set3 = set(csmo['assay_id'])
set4 = set(csgemo['assay_id'])

dict_set = {'csge': set1, 'gemo':set2, 'csmo':set3, 'csgemo':set4}
cmaps = ["cool", list("rgb"), "plasma", "viridis", "Set1"]
plt.figure(figsize=(9,9))
cmap = ["#F8766D","#7CAE00","#C77CFF","#00BFC4"]
venn(dict_set, cmap=cmap, fontsize=16, legend_loc=None)
plt.savefig('../plots/venn_combined_modalities_90.svg')

# Generate Venn diagram for combined modalities and AUC 0.7
df_dataset = pd.read_csv('../predictions/scaffold_median_AUC.csv')
csge = df_dataset[(df_dataset['descriptor'] == 'late_fusion_cs_ge')][(df_dataset['auc_70'] == True)]
gemo = df_dataset[(df_dataset['descriptor'] == 'late_fusion_ge_mobc')][(df_dataset['auc_70'] == True)]
csmo = df_dataset[(df_dataset['descriptor'] == 'late_fusion_cs_mobc')][(df_dataset['auc_70'] == True)]
csgemo = df_dataset[(df_dataset['descriptor'] == 'late_fusion_cs_ge_mobc')][(df_dataset['auc_70'] == True)] 

plt.figure(figsize=(4,4))
set1 = set(csge['assay_id'])
set2 = set(gemo['assay_id'])
set3 = set(csmo['assay_id'])
set4 = set(csgemo['assay_id'])

dict_set = {'csge': set1, 'gemo':set2, 'csmo':set3, 'csgemo':set4}
cmaps = ["cool", list("rgb"), "plasma", "viridis", "Set1"]
plt.figure(figsize=(9,9))
cmap = ["#F8766D","#7CAE00","#C77CFF","#00BFC4"]
venn(dict_set, cmap=cmap, fontsize=16, legend_loc=None)
plt.savefig('../plots/venn_combined_modalities_70.svg')

# Generate Venn diagram for single modalities and AUC 0.7
df_dataset = pd.read_csv('../predictions/scaffold_median_AUC.csv')
cs = df_dataset[(df_dataset['descriptor'] == 'cp_es_op')][(df_dataset['auc_70'] == True)]
ge = df_dataset[(df_dataset['descriptor'] == 'ge_es_op')][(df_dataset['auc_70'] == True)]
mo = df_dataset[(df_dataset['descriptor'] == 'mobc_es_op')][(df_dataset['auc_70'] == True)] 

plt.figure(figsize=(4,4))
set1 = set(cs['assay_id'])
set2 = set(ge['assay_id'])
set3 = set(mo['assay_id'])


# Adjust circle sizes in Illustrator
plt.figure(figsize=(9,9))
dict_set = {'CS': set1, 'GE':set2, 'MO-BC':set3}
cmaps = ["cool", list("rgb"), "plasma", "viridis", "Set1"]
plt.figure(figsize=(9,9))
cmap = ["#F1C232","#6FA8DC","#93C47C", "#d5a6bd"]
venn(dict_set, cmap=cmap, fontsize=16, legend_loc=None)
plt.savefig("../plots/venn_single_modalities_70.svg")
