==> analysis/10-AUC_comparison.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sb

df = pd.read_csv("../predictions/scaffold_median_AUC.csv")

CS = df[df["descriptor"] == "cp_es_op"]
MO = df[df["descriptor"] == "mobc_es_op"]
GE = df[df["descriptor"] == "ge_es_op"]

# x=CS , y=MO
plt.figure(figsize=(6,6))
data = pd.merge(CS, MO, on="assay_id")[["assay_id", "auc_x", "auc_y"]]
data["best"] = data["auc_x"] > data["auc_y"]
print("x > y:", np.sum(data["best"]), "x < y:", data.shape[0] - np.sum(data["best"]))
g = sb.jointplot(data=data, x="auc_x", y="auc_y", kind="scatter", hue="best", 
                 legend=False, xlim=(-.05,1.1), ylim=(-0.05,1.1), palette=["#599b3e", "#ecb727"])

# x=CS , y=GE
plt.figure(figsize=(6,6))
data = pd.merge(CS, GE, on="assay_id")[["assay_id", "auc_x", "auc_y"]]
data["best"] = data["auc_x"] > data["auc_y"]
print("x > y:", np.sum(data["best"]), "x < y:", data.shape[0] - np.sum(data["best"]))
g = sb.jointplot(data=data, x="auc_x", y="auc_y", kind="scatter", hue="best", 
                 legend=False, xlim=(-.05,1.1), ylim=(-0.05,1.1), palette=["#3b70e2", "#ecb727"])

# x=MO , y=GE
plt.figure(figsize=(6,6))
data = pd.merge(MO, GE, on="assay_id")[["assay_id", "auc_x", "auc_y"]]
data["best"] = data["auc_x"] > data["auc_y"]
print("x > y:", np.sum(data["best"]), "x < y:", data.shape[0] - np.sum(data["best"]))
g = sb.jointplot(data=data, x="auc_x", y="auc_y", kind="scatter", hue="best", 
                 legend=False, xlim=(-.05,1.1), ylim=(-0.05,1.1), palette=["#3b70e2", "#599b3e"])

counts = []
for t in np.arange(0.1, 1., 0.01):
    counts.append({"Threshold":t, "Count":np.sum(CS["auc"] > t), "Type":"CS"})
    counts.append({"Threshold":t, "Count":np.sum(GE["auc"] > t), "Type":"GE"})
    counts.append({"Threshold":t, "Count":np.sum(MO["auc"] > t), "Type":"MO"})

plt.gcf().set_size_inches(12, 8)    
counts = pd.DataFrame(data=counts)
sb.lineplot(data=counts, x="Threshold", y="Count", hue="Type", 
            palette=["#ecb727", "#3b70e2", "#599b3e"], linewidth=2)
plt.plot([0.9,0.9],[0.0,372], "--b") # , palette=["#ecb727", "#599b3e", "#3b70e2"]
plt.plot([0.7,0.7],[0.0,372], "--g")
plt.plot([0.5,0.5],[0.0,372], "--r")
plt.savefig('../plots/auc.svg')
