==> analysis/12-Boxplots.ipynb <==
import pandas as pd
import seaborn as sb
import matplotlib.pyplot as plt

modalities = ['mobc_es_op', 'cp_es_op', 'ge_es_op', 'late_fusion_cs_ge', 'late_fusion_cs_ge_mobc', 'late_fusion_cs_mobc', 'late_fusion_ge_mobc']
modalities_dict = {'mobc_es_op':'MO', 'cp_es_op':'CS', 'ge_es_op':'GE', 'late_fusion_cs_ge':'CS+GE', 'late_fusion_cs_ge_mobc':'CS+GE+MO', 'late_fusion_cs_mobc':'CS+MO', 'late_fusion_ge_mobc':'GE+MO'}
colors_dict = {'MO':'#6aa84f', 'CS':'#f1c232', 'GE':'#4a86e8', 'CS+GE':'#f8766d', 'CS+GE+MO':'#00BFC4', 'CS+MO':'#C77CFF', 'GE+MO':'#7cae00'}

auc_file_median = '../predictions/scaffold_median_AUC.csv'
auc_file_median_df = pd.read_csv(auc_file_median)
auc_file_median_df = auc_file_median_df[auc_file_median_df['descriptor'].isin(modalities)]
auc_file_median_df['descriptor'].replace(modalities_dict, inplace=True)

sb.set(rc={'figure.figsize':(11.7,8.27)})
sb.set_style("whitegrid")
sb.boxplot(x = 'descriptor', y="auc", data=auc_file_median_df[auc_file_median_df.descriptor.isin(['CS', 'GE', 'MO'])], palette = colors_dict)
ax = sb.swarmplot(x='descriptor', y="auc", data=auc_file_median_df[auc_file_median_df.descriptor.isin(['CS', 'GE', 'MO'])], palette = colors_dict)
plt.savefig('../plots/single_modalities_box.svg')

sb.set(rc={'figure.figsize':(11.7,8.27)})
sb.set_style("whitegrid")
sb.set_theme(style="ticks")
sb.boxplot(x = 'auc', y="descriptor", data=auc_file_median_df[auc_file_median_df.descriptor.isin(['CS+GE', 'CS+MO', 'GE+MO', 'CS+GE+MO'])], palette = colors_dict, order=['CS+GE', 'CS+MO', 'GE+MO','CS+GE+MO'], width = 0.2, showfliers = False)
ax = sb.stripplot(x='auc', y="descriptor", data=auc_file_median_df[auc_file_median_df.descriptor.isin(['CS+GE', 'CS+MO', 'GE+MO', 'CS+GE+MO'])], palette = colors_dict, order = ['CS+GE', 'CS+MO', 'GE+MO', 'CS+GE+MO'],  linewidth=0)
ax.set_box_aspect(3/len(ax.patches))
ax.xaxis.grid(True)
plt.savefig('../plots/late_fusion_box.svg')
